AbstractCreateJobDispatchedEventsTest is a functional test only
